@@ -11,18 +11,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.3",
+    "@angular/animations": "^8.1.1",
     "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.0.0",
-    "@angular/compiler": "7.0.0",
-    "@angular/core": "^7.2.0",
+    "@angular/common": "^8.1.1",
+    "@angular/compiler": "^8.1.1",
+    "@angular/core": "^8.1.1",
     "@angular/flex-layout": "^7.0.0-beta.24",
-    "@angular/forms": "7.0.0",
+    "@angular/forms": "^8.1.1",
     "@angular/http": "7.1.3",
     "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "7.0.0",
-    "@angular/platform-browser-dynamic": "7.0.0",
-    "@angular/router": "7.0.0",
+    "@angular/platform-browser": "^8.1.1",
+    "@angular/platform-browser-dynamic": "^8.1.1",
+    "@angular/router": "^8.1.1",
     "@kolkov/ngx-dadata": "^0.4.1",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -33,9 +33,9 @@
     "jqwidgets-scripts": "6.2.0",
     "materialize-css": "^1.0.0-rc.2",
     "ng5-slider": "^1.1.13",
-    "rxjs": "6.3.3",
+    "rxjs": "^6.5.2",
     "ymaps": "^0.1.1",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
     "@angular/compiler-cli": "~7.0.0",
@@ -57,5 +57,7 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0"
-  }
+  },
+  "bootstrapVersion": 0,
+  "styleLoaders": []
 }
