@@ -1,3 +1,4 @@
+// @ts-ignore
 import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
@@ -14,7 +15,6 @@ import {RegisterPageComponent} from './auth/register-page/register-page.componen
 import {TokenInterceptor} from './shared/classes/token.interceptor';
 import {NotFoundComponent} from './shared/components/not-found/not-found.component';
 import {MaterialModule} from './shared/material-module';
-// import {NgxDadataModule} from '@kolkov/ngx-dadata';
 
 
 @NgModule({
@@ -24,8 +24,6 @@ import {MaterialModule} from './shared/material-module';
     AuthLayoutComponent,
     RegisterPageComponent,
     NotFoundComponent,
-
-
   ],
   imports: [
     BrowserModule,
@@ -33,7 +31,6 @@ import {MaterialModule} from './shared/material-module';
     ReactiveFormsModule,
     HttpClientModule,
     BrowserAnimationsModule,
-    // NgxDadataModule,
 
     AppRoutingModule,
     SharedModule,
@@ -46,6 +43,9 @@ import {MaterialModule} from './shared/material-module';
       multi: true,
       useClass: TokenInterceptor
     }
+  ],
+  exports: [
+
   ],
   bootstrap: [AppComponent]
 })
