@@ -1,9 +1,10 @@
-import {Component, OnDestroy, OnInit} from '@angular/core'
-import {FormControl, FormGroup, Validators} from '@angular/forms'
-import {AuthService} from '../../shared/services/auth.service'
-import {Router} from '@angular/router'
-import {Subscription} from 'rxjs'
-import {MaterialService} from '../../shared/classes/material.service'
+// @ts-ignore
+import {Component, OnDestroy, OnInit} from '@angular/core';
+import {FormControl, FormGroup, Validators} from '@angular/forms';
+import {AuthService} from '../../shared/services/auth.service';
+import {Router} from '@angular/router';
+import {Subscription} from 'rxjs';
+import {MaterializeService} from '../../shared/classes/materialize.service';
 
 @Component({
   selector: 'app-register-page',
@@ -12,8 +13,8 @@ import {MaterialService} from '../../shared/classes/material.service'
 })
 export class RegisterPageComponent implements OnInit, OnDestroy {
 
-  form: FormGroup
-  aSub: Subscription
+  form: FormGroup;
+  aSub: Subscription;
 
   constructor(private auth: AuthService,
               private router: Router) {
@@ -23,30 +24,30 @@ export class RegisterPageComponent implements OnInit, OnDestroy {
     this.form = new FormGroup({
       email: new FormControl(null, [Validators.required, Validators.email]),
       password: new FormControl(null, [Validators.required, Validators.minLength(6)])
-    })
+    });
   }
 
   ngOnDestroy() {
     if (this.aSub) {
-      this.aSub.unsubscribe()
+      this.aSub.unsubscribe();
     }
   }
 
   onSubmit() {
-    this.form.disable()
+    this.form.disable();
     this.aSub = this.auth.register(this.form.value).subscribe(
       () => {
         this.router.navigate(['/login'], {
           queryParams: {
             registered: true
           }
-        })
+        });
       },
       error => {
-        MaterialService.toast(error.error.message)
-        this.form.enable()
+        MaterializeService.toast(error.error.message);
+        this.form.enable();
       }
-    )
+    );
   }
 
 }
