@@ -1,6 +1,7 @@
+// @ts-ignore
+import {Injectable} from '@angular/core';
 import {ActivatedRouteSnapshot, CanActivate, CanActivateChild, Router, RouterStateSnapshot} from '@angular/router';
 import {Observable, of} from 'rxjs';
-import {Injectable} from '@angular/core';
 import {AuthService} from '../services/auth.service';
 
 @Injectable({
