@@ -1,6 +1,6 @@
-import {ElementRef} from '@angular/core'
+import {ElementRef} from '@angular/core';
 
-declare var M
+declare var M;
 declare var jquery: any;
 declare var $: any;
 
@@ -16,25 +16,26 @@ export interface MaterialDatepicker extends MaterialInstance {
   date?: Date
 }
 
-export class MaterialService {
+export class MaterializeService {
+
   static toast(message: string) {
-    M.toast({html: message})
+    M.toast({html: message});
   }
 
   static initializeFloatingButton(ref: ElementRef) {
-    M.FloatingActionButton.init(ref.nativeElement)
+    M.FloatingActionButton.init(ref.nativeElement);
   }
 
   static updateTextInputs() {
-    M.updateTextFields()
+    M.updateTextFields();
   }
 
   static initModal(ref: ElementRef): MaterialInstance {
-    return M.Modal.init(ref.nativeElement)
+    return M.Modal.init(ref.nativeElement);
   }
 
   static initTooltip(ref: ElementRef): MaterialInstance {
-    return M.Tooltip.init(ref.nativeElement)
+    return M.Tooltip.init(ref.nativeElement);
   }
 
   static initDatepicker(ref: ElementRef, onClose: () => void): MaterialDatepicker {
@@ -42,11 +43,11 @@ export class MaterialService {
       format: 'dd.mm.yyyy',
       showClearBtn: true,
       onClose
-    })
+    });
   }
 
   static initSelect(ref: ElementRef): MaterialInstance {
-    return M.FormSelect.getInstance(ref.nativeElement)
+    return M.FormSelect.getInstance(ref.nativeElement);
   }
 
   // export function initDropdown() {
@@ -57,11 +58,11 @@ export class MaterialService {
 
   static initDropdown() {
     $(document).ready(function () {
-      $(".dropdown-trigger").dropdown();
+      $('.dropdown-trigger').dropdown();
     });
   }
 
   static initSidenav(ref: ElementRef): MaterialInstance {
-    return M.Sidenav.init(ref.nativeElement)
+    return M.Sidenav.init(ref.nativeElement);
   }
 }
