@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+// @ts-ignore
+import {Component, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-auth-layout',
@@ -7,7 +8,8 @@ import { Component, OnInit } from '@angular/core';
 })
 export class AuthLayoutComponent implements OnInit {
 
-  constructor() { }
+  constructor() {
+  }
 
   ngOnInit() {
   }
