@@ -1,3 +1,4 @@
+// @ts-ignore
 import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
@@ -7,9 +8,9 @@ import {SharedModule} from '../../../shared.module';
 import {HandbookLayoutComponent} from './handbook-layout.component';
 import {HandbookRoutingModule} from './handbook-routing.module';
 import {EquipmentTypeComponent} from '../../../../equipment-type/equipment-type.component';
-import {SimpleHandbookComponent} from '../../../components/simple-handbook/simple-handbook.component';
 import {ContragentComponent} from '../../../../contragent/contragent.component';
 import {ContractComponent} from '../../../../contract/contract.component';
+import {JsonFormatedPipe} from '../../../pipe/json-formated.pipe';
 
 
 @NgModule({
@@ -18,8 +19,7 @@ import {ContractComponent} from '../../../../contract/contract.component';
     EquipmentTypeComponent,
     ContragentComponent,
     ContractComponent,
-    SimpleHandbookComponent,
-
+    JsonFormatedPipe,
   ],
   exports: [
     HandbookLayoutComponent
@@ -31,7 +31,7 @@ import {ContractComponent} from '../../../../contract/contract.component';
     HttpClientModule,
 
     HandbookRoutingModule,
-    SharedModule
+    SharedModule,
   ]
 })
 export class HandbookModule {
